requirements.txt
absl-py==2.1.0
aiohappyeyeballs==2.4.3
aiohttp==3.11.4
aiosignal==1.3.1
astunparse==1.6.3
async-timeout==5.0.1
attrs==24.2.0
certifi==2024.8.30
charset-normalizer==3.4.0
datasets==3.1.0
dill==0.3.8
ffmpeg-python==0.2.0
filelock==3.16.1
flatbuffers==24.3.25
frozenlist==1.5.0
fsspec==2024.9.0
future==1.0.0
gast==0.6.0
google-pasta==0.2.0
grpcio==1.68.0
h5py==3.12.1
huggingface-hub==0.26.2
idna==3.10
keras==3.6.0
libclang==18.1.1
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
ml-dtypes==0.3.2
multidict==6.1.0
multiprocess==0.70.16
namex==0.0.8
numpy==1.26.4
opt_einsum==3.4.0
optree==0.13.1
packaging==24.2
pandas==2.2.3
propcache==0.2.0
protobuf==4.25.5
pyarrow==18.0.0
pydub==0.25.1
Pygments==2.18.0
python-dateutil==2.9.0.post0
pytz==2024.2
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.3
rich==13.9.4
safetensors==0.4.5
six @ file:///tmp/build/80754af9/six_1644875935023/work
tensorboard==2.16.2
tensorboard-data-server==0.7.2
tensorflow==2.16.2
tensorflow-io-gcs-filesystem==0.37.1
tensorflow-macos==2.16.2
tensorflow-metal==1.1.0
termcolor==2.5.0
tokenizers==0.20.3
tqdm==4.67.0
transformers==4.46.3
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.2.3
Werkzeug==3.1.3
wrapt==1.16.0
xxhash==3.5.0
yarl==1.17.2
